Fix #2: support nested generic notation like Array<Array<number>>.
The issue was that >> was parsed as a single token '>>' instead
of two tokens '>'. Since we don't ever need to parse a true right
shift '>>', the solution was simply to remove the definition of
'>>' as a token.

In the process, I removed all the operators that are not actually
used.